Add parse/Create wireguard config and a rudimentary version of wg-quick.
I'm publishing this initial version of wg-quick to just execute the `qg-quick` command, but in the next commits it will be introduced in javascript reducing some commands executed.

I'm adding an almost perfect version to build and convert from wg config to json:
  * Some features are not working like `preUp`, `preDown`, `MTU` and `SaveConfig`.

the `wg` command still being thought of as being built!